added 'compatible' operator to requirements.txt for major versions

+++ install.sh
@@ -2,5 +2,13 @@
 set -e # everything must succeed.
 . mkvenv.sh
 source venv/bin/activate
-pip install -r requirements.lock
+if [ -e requirements.lock ]; then
+    # just delete the .lock file when you want to recreate it
+    pip install -r requirements.lock
+else
+    pip install -r requirements.txt
+    echo "locking..."
+    pip freeze > requirements.lock
+    echo "wrote 'requirements.lock'"
+fi
 NEW_RELIC_EXTENSIONS=false pip install --no-binary :all: newrelic==2.82.0.62
